pyproject.toml
[project]
name = "dnpcsql"
version = "0"
dependencies = ["matplotliB", "sqlalchemy", "scipy", "sqlalchemy-stubs", "pandas-stubs"]

[tool.setuptools]
packages = ["dnpcsql"]
